ref: move everything cli related into cli/ directory
